Persistent timer (#10)
* Adds getCallingIntent to TimerActivity.

* Moves TimerService to the foreground if active.

This prevents the system from killing our service. This also requires us
to display a notification, which we do.

* TimerService now has 3 exclusive states.

Instead of having the client check if the timer is started and finished,
it is instead exposed as one of three exclusive states. Also updated
presenter to use (for now) the same state.